SmartGymUser actor added
